refactor existing components
