+++ src/app/shared/components/template/components/button.ts
@@ -3,9 +3,7 @@ import { TemplateBaseComponent } from "./base";
 
 @Component({
   selector: "plh-tmpl-button",
-  template: `<ion-button (click)="triggerActions()">{{
-    _row.value | localVarsReplace: parent.localVariables
-  }}</ion-button>`,
+  template: `<ion-button (click)="triggerActions()">{{ _row.value }}</ion-button>`,
   styleUrls: ["./tmpl-components-common.scss"],
 })
 export class TmplButtonComponent extends TemplateBaseComponent {}
