+++ src/app/shared/components/template/components/layout/animated_section.ts
@@ -1,0 +1,14 @@
+import { Component } from "@angular/core";
+import { TemplateBaseComponent } from "../base";
+
+@Component({
+  selector: "plh-tmpl-animated-section",
+  template: `<div class="animated-section">
+    <plh-template-component
+      *ngFor="let childRow of _row.rows"
+      [row]="childRow"
+      [parent]="parent"
+    ></plh-template-component>
+  </div>`,
+})
+export class AnimatedSectionComponent extends TemplateBaseComponent {}
