+++ src/app/shared/components/template/components/text.ts
@@ -3,7 +3,7 @@ import { TemplateBaseComponent } from "./base";
 
 @Component({
   selector: "plh-tmpl-text",
-  template: `<p>{{ _row.value | localVarsReplace: parent.localVariables }}</p>`,
+  template: `<p>{{ _row.value }}</p>`,
   styleUrls: ["./tmpl-components-common.scss"],
 })
 export class TmplTextComponent extends TemplateBaseComponent {}
