+++ src/app/shared/components/template/components/timer.ts
@@ -1,19 +1,9 @@
-import { Component, Input, OnInit } from "@angular/core";
-import { FlowTypes } from 'src/app/shared/model/flowTypes';
-import { ITemplateComponent } from "./tmpl.component";
+import { Component } from "@angular/core";
+import { TemplateBaseComponent } from "./base";
 
 @Component({
   selector: "plh-tmpl-timer",
   template: `<p>Timer to be implemented</p>`,
-  styleUrls: ["./tmpl-components-common.scss"]
+  styleUrls: ["./tmpl-components-common.scss"],
 })
-export class TmplTimerComponent implements ITemplateComponent, OnInit {
-
-  @Input() row: FlowTypes.TemplateRow;
-  @Input() template: FlowTypes.Template;
-  @Input() localVariables: { [name: string]: string};
-
-  constructor() {}
-
-  ngOnInit() {}
-}
+export class TmplTimerComponent extends TemplateBaseComponent {}
