+++ src/app/shared/components/template/components/title.ts
@@ -3,7 +3,7 @@ import { TemplateBaseComponent } from "./base";
 
 @Component({
   selector: "plh-tmpl-title",
-  template: ` <h1>{{ _row.value | localVarsReplace: _localVariables }}</h1> `,
+  template: ` <h1>{{ _row.value }}</h1> `,
   styleUrls: ["./tmpl-components-common.scss"],
 })
 export class TmplTitleComponent extends TemplateBaseComponent {}
